Created and tested Star Wars character DOB function
Created, type hinted, and tested a Star Wars character DOB function
using the SWAPI api. The function searches for DOB using the name
(or substring of name) and returns the DOB if name is found. I tested
the function including mocking the exceptions and requests.get. I had
to add a pragma to the next call on the generator expression as
coverage though that variable prior to exit immediately. Seems to be a
bug related to this issue
https://github.com/nedbat/coveragepy/issues/515.